requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
Django==5.0.2
djangorestframework==3.14.0
MouseInfo==0.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
